+++ test/eslint.config.json.js
@@ -1,6 +1,6 @@
 import globals from "globals";
 import pluginJs from "@eslint/js";
-import jsonSchemaDescriptions from "../dist/index.js";
+import jsonSchemaDescriptions from "../dist/recommended/index.js";
 
 export default [
   { languageOptions: { globals: globals.node } },